Fix batches/annotations route
The get_latest_taxonomy_ids() function returned the ids of
Objects and Landcover classes from the Taxonomy db. That id
was then used in a query on the TaxonomyClass db, where the
Taxonomy id of the Landcover class did not match with it's
TaxonomyClass id (not to be confused with taxonomy_id, which
is a foreign key in that db).